migration updated message class and controller added

+++ app/Http/Controllers/Profile/OfferController.php
@@ -93,7 +93,7 @@ class OfferController extends Controller
             $user = Auth::user();
             return view('profile.offers.edit', compact('user', 'offer'));
         }
-        return redirect()->back()->with('warning', array('This offer can not be edited'=>'Once an offer has been accepted it cannot be edited anymore!'));
+        return redirect()->back()->with('warning', array('This offer can not be edited'=>'Once an offer has been accepted it cannot be edited anymore!  Request the orderer to remove accepted offer to update it.'));
     }
 
     /**
@@ -121,7 +121,10 @@ class OfferController extends Controller
     public function destroy($id)
     {
         $offer = $this->offer->findOrFail($id);
-        $offer->delete();
-        return redirect()->route('offers.index')->with('success', array('Success'=>'Offer has been deleted!'));
+        if(empty($offer->accepted)){
+            $offer->delete();
+            return redirect()->route('offers.index')->with('success', array('Success'=>'Offer has been deleted!'));
+        }
+        return redirect()->back()->with('warning', array('This offer can not be deleted'=>'Once an offer has been accepted it cannot be deleted anymore! Request the orderer to remove accepted offer to delete it.'));
     }
 }
